Reconstruye modulos de Lotes, Ordenes de Compra y Movimientos de Inventario
- Lotes: alinea dominio/casos de uso/DTOs con el controller y adaptador
  existentes (estaban a medio reescribir y no compilaban).
- Ordenes de Compra: implementa capa REST completa (repositorio, caso de
  uso con maquina de estados BORRADOR/ENVIADA/RECIBIDA/ANULADA, DTOs,
  mapper, controller) adaptando la entidad al contrato simple que ya
  espera el front2 (codigoReferencia, proveedorId, detalles, version).
- Movimientos de Inventario: reconstruye el modulo completo (antes solo
  tenia codigo/cantidad/precio) para soportar tipo, producto, lote,
  descripcion y paginacion (Page) tal como lo consume el front2.
- Corrige bugs de tipo Integer/Long en varios JpaRepository (Lote,
  Movimiento) y agrega findByCodigo a EstadoOc/TipoMovimiento.
- Ajusta DrinkHouseConfig con el wiring correcto de todos los beans
  nuevos/modificados.

No se pudo compilar en este entorno (sin JDK/Maven); revisado
manualmente con cuidado. Recomendado correr 'mvnw clean compile' antes
de fusionar.

+++ src/main/java/com/uisrael/drinkhouse/infraestructura/repositorio/IMovimientoInventarioJpaRepositorio.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.uisrael.drinkhouse.infraestructura.persistencia.jpa.MovimientoInventarioEntity;
 
-public interface IMovimientoInventarioJpaRepositorio extends JpaRepository<MovimientoInventarioEntity, Integer> {
+public interface IMovimientoInventarioJpaRepositorio extends JpaRepository<MovimientoInventarioEntity, Long> {
 
 }
